refactor(theme): nest theme icons under subfolder @theme/Icon/* (#7740)
