+++ packages/docusaurus-theme-classic/src/theme/Icon/DarkMode/index.tsx
@@ -6,7 +6,7 @@
  */
 
 import React from 'react';
-import type {Props} from '@theme/IconDarkMode';
+import type {Props} from '@theme/Icon/DarkMode';
 
 export default function IconDarkMode(props: Props): JSX.Element {
   return (
