+++ packages/docusaurus-theme-classic/src/theme/Icon/Edit/index.tsx
@@ -7,7 +7,7 @@
 
 import React from 'react';
 import clsx from 'clsx';
-import type {Props} from '@theme/IconEdit';
+import type {Props} from '@theme/Icon/Edit';
 
 import styles from './styles.module.css';
