+++ packages/docusaurus-theme-classic/src/theme/Icon/Edit/styles.module.css
