+++ packages/docusaurus-theme-classic/src/theme/Icon/ExternalLink/styles.module.css
