+++ packages/docusaurus-theme-classic/src/theme/Icon/Menu/index.tsx
@@ -6,7 +6,7 @@
  */
 
 import React from 'react';
-import type {Props} from '@theme/IconMenu';
+import type {Props} from '@theme/Icon/Menu';
 
 export default function IconMenu({
   width = 30,
