+++ packages/docusaurus-theme-classic/src/theme/Navbar/MobileSidebar/Header/index.tsx
@@ -8,7 +8,7 @@
 import React from 'react';
 import {useNavbarMobileSidebar} from '@docusaurus/theme-common/internal';
 import NavbarColorModeToggle from '@theme/Navbar/ColorModeToggle';
-import IconClose from '@theme/IconClose';
+import IconClose from '@theme/Icon/Close';
 import NavbarLogo from '@theme/Navbar/Logo';
 
 function CloseButton() {
