+++ packages/docusaurus-theme-classic/src/theme/Navbar/MobileSidebar/Toggle/index.tsx
@@ -7,7 +7,7 @@
 
 import React from 'react';
 import {useNavbarMobileSidebar} from '@docusaurus/theme-common/internal';
-import IconMenu from '@theme/IconMenu';
+import IconMenu from '@theme/Icon/Menu';
 
 export default function MobileSidebarToggle(): JSX.Element {
   const mobileSidebar = useNavbarMobileSidebar();
