+++ website/src/plugins/changelog/theme/types.d.ts
@@ -15,7 +15,7 @@ declare module '@theme/ChangelogItem/Header/Authors';
 declare module '@theme/ChangelogList';
 declare module '@theme/ChangelogList/Header';
 
-declare module '@theme/IconExpand' {
+declare module '@theme/Icon/Expand' {
   import type {ComponentProps} from 'react';
 
   export interface Props extends ComponentProps<'svg'> {
